+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 1.0.0
-
 * Initial Release
+
+## 1.1.0
+Bugix:
+* Disposing Listener throwed because of depricated context 
+Feature:
+* Added Loading Flag in UserStateService